collision_moving_pawn is an iterator, bugfixes.
collision_moving_pawn is now an iterator.

also, collision_moving_pawn is just plain smarter now.
if movement in x and movement in y got hits at the exact same time t,
it yields the union of those results, rather than yielding twice
with the same time t.

also, found that bug I mentioned in the previous comment.
fixed, and added regression test.  the bug was a malformed
vector I used in the code:
    vec2_1_0 = vec2(1, 1)
or something like that.